Move graphToPrompt to executionUtil
